Add vim-slime for interactive command usage
The typical usage of this plugin is to use vim for entering commands
which will be receive by an REPL in another screen or pane.

Add a toggle for IPython because of the indentation issues when using
IPython. IPython automatically put an indentation per line so pasting
a code will mess up the indentation of the code leading to syntax
errors.