Remove minibufexpl because of fugitive conflict
When invoking :Gdiff, minibufexpl has a bug which interfere with the
functionality of the command. Instead of showing the diff of the changes, the
diff shown is between the tab bar of minibufexpl and the buffer.

It is a hassle and inconvinient because of the frequency of the usage of Gdiff
so remove the plugin.