Make use of ncmpcpp instead of cmus
This player has more features than cmus. It has its own visualizer and
the fact that it uses mpd makes it much better. Though it requires relearning
many keybindings. It can fetch lyrics on its own too.

The player will be in a window composed of three panes. Those panes will
composed of visualizer, lyrics and the current playlist.