Make tmux_session an sh script and add ncmpcpp
The layout for ncmpcpp is complicated. At the top, there are two panes. At the
left is the visualizer and at the right is the lyrics for the current song.
Because there is no shortcut for lyrics view in the command line, tmux buffer
is needed. The problem is that the execution of commands is so fast that the
layout got wrong. Or, they are executed in the background. To make it more
smooth, convert the file instead to an sh script. Problem is solved.