Use Ctrl-G as the prefix key and fix send prefix
Alt-a is really hard to work on variety of devices that I had tried.
Some devices does not react when those combination is pressed. Rather
than bearing these inconvieniences when moving from one to another,
change the prefix key instead.

This also enabled sending of the prefix key which before was overriden by
the last window setting.